这个项目停了

+++ equipment-java/equipmanager-module-oam/equipmanager-module-oam-biz/src/main/java/cn/newness/imip/module/oam/service/maintaindetail/MaintainDetailServiceImpl.java
@@ -1,6 +1,6 @@
 package cn.newness.imip.module.oam.service.maintaindetail;
 
-import cn.newness.imip.framework.mybatis.core.query.LambdaQueryWrapperX;
+import cn.newness.imip.module.property.api.EquipApi;
 import org.redisson.api.RLock;
 import org.redisson.api.RedissonClient;
 import org.springframework.stereotype.Service;
@@ -29,6 +29,8 @@ public class MaintainDetailServiceImpl implements MaintainDetailService {
     private MaintainDetailMapper maintainDetailMapper;
     @Resource
     private RedissonClient redissonClient;//对于条件修改的加锁
+    @Resource
+    private EquipApi equipApi;
 
     @Override
     public String createMaintainDetail(MaintainDetailSaveReqVO createReqVO) {
@@ -36,25 +38,15 @@ public class MaintainDetailServiceImpl implements MaintainDetailService {
         lock.lock();
         try{
             createReqVO.setId(BeanUtils.createId());
-            //判断同一设备/特殊设备档案是否已经有了更换自身的保养项？有了就重复了：
-            // 1、非特殊设备重复；[equipId相同，isSpecial都为0]
-            // 2、特殊设备设备档案重复[equipId相同，isSpecial都为1，equipprofileId相同]
-            if(createReqVO.getReplaceSelf()==1){
-                MaintainDetailPageReqVO md=new MaintainDetailPageReqVO();
-                md.setEquipId(createReqVO.getEquipId());
-                md.setReplaceSelf(1);
-                md.setIsSpecial(createReqVO.getIsSpecial());
-                if(createReqVO.getIsSpecial()==1){
-                    md.setEquipprofileId(createReqVO.getEquipprofileId());
-                }
-                int count = getList(md).size();
-                if(count>0){
-                    throw exception(MAINTAIN_MORE_REPLACE_ERROR);
+            if(createReqVO.getIsSpecial()==0){
+                if(createReqVO.getRefertoId() != null && !"".equals(createReqVO.getRefertoId())){
+                    throw exception(MAINTAIN_REFERTO_ERROR);
+                }else {
+                    createReqVO.setIsReferto(0);
                 }
             }
-            if(createReqVO.getIsSpecial()==0 && createReqVO.getRefertoId() != null && !"".equals(createReqVO.getRefertoId())){
-                throw exception(MAINTAIN_REFERTO_ERROR);
-            }
+            //赋值为全层级名
+            createReqVO.setEquipName(equipApi.getAllLayerEquipName(createReqVO.getEquipId()));
             // 插入
             MaintainDetailDO maintainDetail = BeanUtils.toBean(createReqVO, MaintainDetailDO.class);
             maintainDetailMapper.insert(maintainDetail);
@@ -67,29 +59,31 @@ public class MaintainDetailServiceImpl implements MaintainDetailService {
 
     @Override
     public void updateMaintainDetail(MaintainDetailSaveReqVO updateReqVO) {
-        MaintainDetailDO oldMaintainDetail;//之后用作操作日志
+        MaintainDetailDO oldMaintainDetail = validateMaintainDetailExists(updateReqVO.getId());//之后用作操作日志
         RLock lock = redissonClient.getLock("maintainDetail" + updateReqVO.getId());
         lock.lock();
         try{
-            if(updateReqVO.getReplaceSelf()==1){
-                //判断同一设备是否已经有了更换自身的保养项？有了就重复了：1、非特殊设备重复；2、特殊设备设备档案重复
-                MaintainDetailPageReqVO md=new MaintainDetailPageReqVO();
-                md.setEquipId(updateReqVO.getEquipId());
-                md.setReplaceSelf(1);
-                md.setIsSpecial(updateReqVO.getIsSpecial());
-                if(updateReqVO.getIsSpecial()==1){
-                    md.setEquipprofileId(updateReqVO.getEquipprofileId());
+            //先判断该项内容有没有被作为参考对象，有的话那就只允许修改周期，其他均不允许被修改
+            if(maintainDetailMapper.selectList("referto_id",updateReqVO.getId()).size()>0){
+                if(
+                        !updateReqVO.getIsSpecial().equals(oldMaintainDetail.getIsSpecial()) ||
+                        !updateReqVO.getEquipId().equals(oldMaintainDetail.getEquipId()) ||
+                        !updateReqVO.getDetails().equals(oldMaintainDetail.getDetails())
+                ){
+                    throw exception(MAINTAIN_UPDATE_ERROR);
                 }
-                List<MaintainDetailDO> detailDOList = getList(md);
-                for (MaintainDetailDO detail:detailDOList){
-                    if(detail.getId().equals(updateReqVO.getId())){
-                        oldMaintainDetail = detail;
-                    }else {
-                        throw exception(MAINTAIN_MORE_REPLACE_ERROR);
-                    }
+            }
+            //修改时如果是非特殊设备也需要搞全层级设备名，不管是不是定期更换计划
+            if(updateReqVO.getIsSpecial()==0){
+                if(updateReqVO.getRefertoId() != null && !"".equals(updateReqVO.getRefertoId())){
+                    throw exception(MAINTAIN_REFERTO_ERROR);
+                }else {
+                    updateReqVO.setIsReferto(0);
                 }
-            }else {
-                oldMaintainDetail = validateMaintainDetailExists(updateReqVO.getId());
+            }
+            if (!oldMaintainDetail.getEquipId().equals(updateReqVO.getEquipId())) {
+                //赋值为全层级名
+                updateReqVO.setEquipName(equipApi.getAllLayerEquipName(updateReqVO.getEquipId()));
             }
             // 更新
             MaintainDetailDO updateObj = BeanUtils.toBean(updateReqVO, MaintainDetailDO.class);
@@ -103,6 +97,10 @@ public class MaintainDetailServiceImpl implements MaintainDetailService {
     public void deleteMaintainDetail(String id) {
         // 校验存在
         validateMaintainDetailExists(id);
+        //校验是否被作为参考项目
+        if(maintainDetailMapper.selectList("referto_id",id).size()>0){
+            throw exception(MAINTAIN_DELETE_ERROR);
+        }
         // 删除
         maintainDetailMapper.deleteById(id);
     }
@@ -149,21 +147,7 @@ public class MaintainDetailServiceImpl implements MaintainDetailService {
                     maintainDetailRespVO.setIsRefertoName("是");
                 }
             }
-            if(maintainDetailRespVO.getReplaceSelf()==0){//是否更换自身
-                maintainDetailRespVO.setReplaceSelfName("否");
-            }else if(maintainDetailRespVO.getReplaceSelf()==1) {
-                maintainDetailRespVO.setReplaceSelfName("是");
-            }
         });
         return maintainDetailRespVOS;
-    }
-
-    private List<MaintainDetailDO> getList(MaintainDetailPageReqVO md){
-        LambdaQueryWrapperX<MaintainDetailDO> queryWrapperX = new LambdaQueryWrapperX<>();
-        queryWrapperX.eq(MaintainDetailDO::getEquipId, md.getEquipId());
-        queryWrapperX.eq(MaintainDetailDO::getIsSpecial, md.getIsSpecial());
-        queryWrapperX.eqIfPresent(MaintainDetailDO::getEquipprofileId,md.getEquipprofileId());
-        queryWrapperX.eq(MaintainDetailDO::getReplaceSelf,md.getReplaceSelf());
-        return maintainDetailMapper.selectList(queryWrapperX);
     }
 }
